Computing Center of Gravity?
I've worked on the behavior characterization. Getting the max altitude (apogee) was easy. I initially wanted to have the "stability" which apparently depends on the CG (Center of Gravity) and CP (what is this?), but computing these things is not trivial, and has changed a lot between different versions of Open Rocket. Looking at the latest version of the code in Java was not helpful, since orhelper communicates via JPype with version 15.03 (available at https://github.com/openrocket/openrocket/tree/release-15.03 ). However, after looking at that code, I think I've been able to calculate the Center of Gravity. I'm not 100% sure it is correct, since it depends on a so-called "default" configuration from the rocket, but I'm not sure this matches the actual simulation configuration that I'm using. Will need to investigate further.