Random genomes work, except for fin points
This currently generates random rockets from random genomes, and there is quite a lot of variance, even though all models have the exact same fin design. For some reason, randomly generated fin points do not work. I'm not sure if there is some sort of actual syntax or logic error, of if randomly selected points somehow create illegal fin shapes that crash in an exception. Will need to investigate further. Maybe just set one of the fin points randomly to see if it is possible for a random point to ever work.