Can generate random rocket designs
There is still an issue with the fin points. Basically, if the random points cause an intersection, then the code throws an exception. My code catches this exception and simply leaves the rocket with the default fin points (sort of ... need to fix this)

Currently trying to plot scatter plot of stability vs altitude at the end, but it crashes. Need to work on that.