Fix bounds issue
I think this has fixed the bounds issue with the solution genomes, but I need to apply this fix to every type of emitter, not just the Gaussian Emitter used with MAP Elites. However, there is a new error I want to solve first. The rocket simulation runs but then crashes with a Java exception. I expect it is an issue with multithreading, so I'm going to try adding synchronization