Adding code to prevent bad fin points
There are a lot of ways to design bad fins. An obvious problem is if all x or all y coordinates are 0.0. That is easy to check, and doing so prevents errors (default to trapezoidal fins).

There are still some errors, such as when all but one point are (0,0). Need to have a fix for that too, but will first clean up some other debugging code.