Load DB of Body Tubes
This is a fairly big change. The orhelper Python code doesn't give access to the database of preset components, even though it loads the database. Very annoying. I had to change the orhelper source code to make the DB accessible, which meant also removing orhelper as an easily installed Python package. Instead, I have to import and run my modified version of the code. Importantly, this makes all the scripts in "examples" not work unless they are moved to the main directory.

Right now, I'm able to load the list of preset body tubes. Interestingly, the package paths in the jar file don't seem to completely match what I'm seeing in the GitHub repo for version 15.03:
https://github.com/openrocket/openrocket/releases/tag/release-15.03
I had to look in the jar file itself to make sense of what to import