Almost implemented body tubes
I am able to make the genome correspond to a specific body tube preset in the database. Just need to figure out how to load/assign it. Also, I removed the part of the genome that defines the nose cone's aft radius, since that will instead need to match the radius of the body tube.