New binning scheme
This is meant to be a 3D bin: nose type (int index), stability, and altitude. Still uses consistency as fitness.

However, for display purposes in pyribs, the stability and nose type are combined along one dimension. It seems to run fine, but I'll need to wait to see how the results actually turn out and display.