Still screwed up
This looks better, but it is still screwed up. The only way to get it to display right is to use a mesh of some sort I think. The original Pyribs visualizer uses pcolormesh. However, simply asking Claude or ChatGPT to use this instead is not creating satisfying results either. Need to investigate more.